Cleanup top-level modules
- All chunk modules become submodules of `chunk`.
- Modules for reading foreign formats become submodules of `compat`.
