+++ src/chunks/mod.rs
@@ -1,0 +1,11 @@
+//! finalfusion chunks
+
+pub(crate) mod io;
+
+pub mod metadata;
+
+pub mod norms;
+
+pub mod storage;
+
+pub mod vocab;
