+++ src/compat/fasttext/indexer.rs
