+++ src/compat/fasttext/mod.rs
