+++ src/compat/mod.rs
@@ -1,0 +1,7 @@
+//! Readers/writers for other embedding formats.
+
+pub mod fasttext;
+
+pub mod text;
+
+pub mod word2vec;
